new changes

+++ src/main/java/com/agile/agile_project/service/Impl/EpicServiceImpl.java
@@ -1,9 +1,11 @@
 package com.agile.agile_project.service.Impl;
 
+import com.agile.agile_project.exception.EpicNotFoundException;
 import com.agile.agile_project.model.Epic;
 import com.agile.agile_project.repository.EpicRepository;
 import com.agile.agile_project.service.EpicService;
 import org.springframework.stereotype.Service;
+
 import java.util.List;
 
 @Service
@@ -23,7 +25,7 @@ public class EpicServiceImpl implements EpicService {
     @Override
     public Epic getById(Long id) {
         return epicRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Epic not found"));
+                .orElseThrow(() -> new EpicNotFoundException(id));
     }
 
     @Override
@@ -33,7 +35,9 @@ public class EpicServiceImpl implements EpicService {
 
     @Override
     public void delete(Long id) {
+        if (!epicRepository.existsById(id)) {
+            throw new EpicNotFoundException(id);
+        }
         epicRepository.deleteById(id);
     }
 }
-
